[storage] move state_value_writer to storage_interface
